Add check for used providers

+++ backend/src/main/java/com/alligator/market/backend/quotes/stream/providers/list/service/ProviderServiceImpl.java
@@ -4,8 +4,10 @@ import com.alligator.market.backend.quotes.stream.providers.list.dto.ProviderCre
 import com.alligator.market.backend.quotes.stream.providers.list.dto.ProviderDto;
 import com.alligator.market.backend.quotes.stream.providers.list.dto.ProviderUpdateDto;
 import com.alligator.market.backend.quotes.stream.providers.list.entity.Provider;
+import com.alligator.market.backend.quotes.stream.ccypair_feed_settings.repository.SettingsRepository;
 import com.alligator.market.backend.quotes.stream.providers.list.exceptions.DuplicateProviderException;
 import com.alligator.market.backend.quotes.stream.providers.list.exceptions.ProviderNotFoundException;
+import com.alligator.market.backend.quotes.stream.providers.list.exceptions.ProviderUsedInSettingsException;
 import com.alligator.market.backend.quotes.stream.providers.list.repository.ProviderRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -25,6 +27,7 @@ import java.util.List;
 public class ProviderServiceImpl implements ProviderService {
 
     private final ProviderRepository repository;
+    private final SettingsRepository settingsRepository;
 
     //==========================
     // Создать нового провайдера
@@ -72,6 +75,11 @@ public class ProviderServiceImpl implements ProviderService {
 
         Provider provider = repository.findByName(name)
                 .orElseThrow(() -> new ProviderNotFoundException(name));
+
+        // Проверка, что провайдер не используется в настройках
+        if (settingsRepository.existsByProvider_Name(name)) {
+            throw new ProviderUsedInSettingsException(name);
+        }
 
         repository.delete(provider);
         log.info("Provider {} deleted (id={})", provider.getName(), provider.getId());
